requirements.txt
python-dotenv==1.01
lxml==5.3.1
python-docx==1.1.2